chore: make fixtures suitable for mssql
- add non-sqlite prompt for initdb, dropdb, load-fixtures
- role column should be of explicit length
- truncate geo hashes

+++ app/commands/database_commands.py
@@ -2,7 +2,7 @@ from typing import Annotated
 
 import click
 from dependency_injector.wiring import Provide, inject
-from sqlalchemy import Engine, event
+from sqlalchemy import Engine, event, text
 from sqlalchemy.orm import Session
 
 from app.api.api_container import ApiContainer
@@ -12,13 +12,11 @@ from tests.fixtures.internal.fixtures_service import FixturesService
 
 @click.command()
 @inject
-def initdb(
-    db_engine: Annotated[Engine, Provide[ApiContainer.db_engine]],
-    database_uri: Annotated[str, Provide[ApiContainer.config.SQLALCHEMY_DATABASE_URI]],
-):
+def initdb(db_engine: Annotated[Engine, Provide[ApiContainer.db_engine]]):
     click.echo("Initialized the database")
-    if database_uri[0:6] != "sqlite":
-        raise RuntimeError("Can only run `initdb` for sqlite")
+    if db_engine.dialect.name != "sqlite":
+        if not click.confirm("Database is not SQLite. Do you want to continue?"):
+            return
 
     table_metadata.create_all(db_engine)
     click.echo("Done")
@@ -26,13 +24,11 @@ def initdb(
 
 @click.command()
 @inject
-def dropdb(
-    db_engine: Annotated[Engine, Provide[ApiContainer.db_engine]],
-    database_uri: Annotated[str, Provide[ApiContainer.config.SQLALCHEMY_DATABASE_URI]],
-):
+def dropdb(db_engine: Annotated[Engine, Provide[ApiContainer.db_engine]]):
     click.echo("Dropping database")
-    if database_uri[0:6] != "sqlite":
-        raise RuntimeError("Can only run `dropdb` for sqlite")
+    if db_engine.dialect.name != "sqlite":
+        if not click.confirm("Database is not SQLite. Do you want to continue?"):
+            return
 
     table_metadata.drop_all(db_engine)
     click.echo("Dropped the database")
@@ -40,22 +36,26 @@ def dropdb(
 
 @click.command()
 @inject
-def load_fixtures(
-    db_engine: Annotated[Engine, Provide[ApiContainer.db_engine]],
-    database_uri: Annotated[str, Provide[ApiContainer.config.SQLALCHEMY_DATABASE_URI]],
-):
+def load_fixtures(db_engine: Annotated[Engine, Provide[ApiContainer.db_engine]]):
     click.echo("Loading fixtures")
-    if database_uri[0:6] != "sqlite":
-        raise RuntimeError("Can only run `load_fixtures` for sqlite")
+    if db_engine.dialect.name != "sqlite":
+        if not click.confirm("Database is not SQLite. Do you want to continue?"):
+            return
 
     def _disable_foreign_keys(dbapi_connection, _record) -> None:
-        dbapi_connection.execute("PRAGMA foreign_keys = OFF")
+        if db_engine.dialect.name == "sqlite":
+            dbapi_connection.execute("PRAGMA foreign_keys = OFF")
+        elif db_engine.dialect.name == "mssql":
+            dbapi_connection.execute("""EXEC sp_MSforeachtable "ALTER TABLE ? NOCHECK CONSTRAINT ALL";""")
 
     event.listen(db_engine, "connect", _disable_foreign_keys)
     try:
         with Session(db_engine) as session:
             FixturesService().load(session)
             session.commit()
+
+            if db_engine.dialect.name == "mssql":
+                session.execute(text("""EXEC sp_MSforeachtable "ALTER TABLE ? WITH CHECK CHECK CONSTRAINT ALL";"""))
     finally:
         event.remove(db_engine, "connect", _disable_foreign_keys)
     click.echo("Done")
